a lot of Board work, some GUI work
changed tile image sizes (will be tiny in the draw pile but will be bigger in a Player hand)

altered a function description in Game

might have to create ITile (a class that will allow us to actually store Tiles in a game board format "underneath" the Board for easier image updating. But we'll see.