AI Actions
I added the code for the AI action in the Board Class. The code is currently commented out as I haven't been able to thoroughly test it yet, and I didn't want to include something so critical if it wasn't fully sure it would work. That said, the code does allow the AI to claim chis, pongs, and kongs. The board does not update their hands however. The comments for AI action were left alone, so feel free to ctrl f to find it.